# docs/smartmed.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SmartMed 本地病历草稿",
    "version": "0.1.0"
  },
  "paths": {
    "/health": {
      "get": {
        "summary": "Health",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/v1/form": {
      "get": {
        "summary": "Form",
        "operationId": "form_v1_form_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/v1/drafts": {
      "post": {
        "summary": "Generate",
        "operationId": "generate_v1_drafts_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExtractRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Draft"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/drafts/{draft_id}": {
      "get": {
        "summary": "Read",
        "operationId": "read_v1_drafts__draft_id__get",
        "parameters": [
          {
            "name": "draft_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Draft Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Draft"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Edit",
        "operationId": "edit_v1_drafts__draft_id__put",
        "parameters": [
          {
            "name": "draft_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Draft Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Draft"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/drafts/{draft_id}/confirm": {
      "post": {
        "summary": "Confirm",
        "operationId": "confirm_v1_drafts__draft_id__confirm_post",
        "parameters": [
          {
            "name": "draft_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Draft Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Draft"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/drafts/{draft_id}/history": {
      "get": {
        "summary": "History",
        "operationId": "history_v1_drafts__draft_id__history_get",
        "parameters": [
          {
            "name": "draft_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Draft Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/drafts/{draft_id}/export": {
      "get": {
        "summary": "Export",
        "operationId": "export_v1_drafts__draft_id__export_get",
        "parameters": [
          {
            "name": "draft_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Draft Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Draft"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ConfirmRequest": {
        "properties": {
          "expected_version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Expected Version"
          },
          "reviewer": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "pattern": "\\S",
            "title": "Reviewer"
          },
          "reviewed": {
            "type": "boolean",
            "const": true,
            "title": "Reviewed"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "expected_version",
          "reviewer",
          "reviewed"
        ],
        "title": "ConfirmRequest"
      },
      "Draft": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Version"
          },
          "state": {
            "type": "string",
            "enum": [
              "requires_review",
              "confirmed"
            ],
            "title": "State"
          },
          "contract_version": {
            "type": "string",
            "title": "Contract Version"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          },
          "turns": {
            "items": {
              "$ref": "#/components/schemas/Turn"
            },
            "type": "array",
            "title": "Turns"
          },
          "document": {
            "$ref": "#/components/schemas/Extraction"
          },
          "model_metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Model Metadata"
          },
          "reviewer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewer"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "version",
          "state",
          "contract_version",
          "created_at",
          "updated_at",
          "turns",
          "document",
          "model_metadata",
          "reviewer"
        ],
        "title": "Draft"
      },
      "EditRequest": {
        "properties": {
          "expected_version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Expected Version"
          },
          "actor": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "pattern": "\\S",
            "title": "Actor"
          },
          "document": {
            "$ref": "#/components/schemas/Extraction"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "expected_version",
          "actor",
          "document"
        ],
        "title": "EditRequest"
      },
      "Evidence": {
        "properties": {
          "turn_id": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Turn Id"
          },
          "quote": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Quote"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "turn_id",
          "quote"
        ],
        "title": "Evidence"
      },
      "ExtractRequest": {
        "properties": {
          "turns": {
            "items": {
              "$ref": "#/components/schemas/Turn"
            },
            "type": "array",
            "maxItems": 40,
            "minItems": 1,
            "title": "Turns"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "turns"
        ],
        "title": "ExtractRequest"
      },
      "Extraction": {
        "properties": {
          "发热": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PresentFact"
              },
              {
                "$ref": "#/components/schemas/NegatedFact"
              },
              {
                "$ref": "#/components/schemas/UnknownFact"
              }
            ],
            "title": "发热"
          },
          "咳嗽": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PresentFact"
              },
              {
                "$ref": "#/components/schemas/NegatedFact"
              },
              {
                "$ref": "#/components/schemas/UnknownFact"
              }
            ],
            "title": "咳嗽"
          },
          "症状持续时间": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PresentFact"
              },
              {
                "$ref": "#/components/schemas/NegatedFact"
              },
              {
                "$ref": "#/components/schemas/UnknownFact"
              }
            ],
            "title": "症状持续时间"
          },
          "青霉素过敏": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PresentFact"
              },
              {
                "$ref": "#/components/schemas/NegatedFact"
              },
              {
                "$ref": "#/components/schemas/UnknownFact"
              }
            ],
            "title": "青霉素过敏"
          },
          "已用药物": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PresentFact"
              },
              {
                "$ref": "#/components/schemas/NegatedFact"
              },
              {
                "$ref": "#/components/schemas/UnknownFact"
              }
            ],
            "title": "已用药物"
          },
          "已用药物剂量": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PresentFact"
              },
              {
                "$ref": "#/components/schemas/NegatedFact"
              },
              {
                "$ref": "#/components/schemas/UnknownFact"
              }
            ],
            "title": "已用药物剂量"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "发热",
          "咳嗽",
          "症状持续时间",
          "青霉素过敏",
          "已用药物",
          "已用药物剂量"
        ],
        "title": "Extraction"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "NegatedFact": {
        "properties": {
          "status": {
            "type": "string",
            "const": "negated",
            "title": "Status"
          },
          "value": {
            "type": "null",
            "title": "Value"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/Evidence"
            },
            "type": "array",
            "maxItems": 8,
            "minItems": 1,
            "title": "Evidence"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "status",
          "value",
          "evidence"
        ],
        "title": "NegatedFact"
      },
      "PresentFact": {
        "properties": {
          "status": {
            "type": "string",
            "const": "present",
            "title": "Status"
          },
          "value": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Value"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/Evidence"
            },
            "type": "array",
            "maxItems": 8,
            "minItems": 1,
            "title": "Evidence"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "status",
          "value",
          "evidence"
        ],
        "title": "PresentFact"
      },
      "Turn": {
        "properties": {
          "turn_id": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Turn Id"
          },
          "speaker": {
            "type": "string",
            "enum": [
              "patient",
              "doctor"
            ],
            "title": "Speaker"
          },
          "text": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Text"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "turn_id",
          "speaker",
          "text"
        ],
        "title": "Turn"
      },
      "UnknownFact": {
        "properties": {
          "status": {
            "type": "string",
            "const": "unknown",
            "title": "Status"
          },
          "value": {
            "type": "null",
            "title": "Value"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/Evidence"
            },
            "type": "array",
            "maxItems": 0,
            "title": "Evidence"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "status",
          "value",
          "evidence"
        ],
        "title": "UnknownFact"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}